requirements.txt
streamlit
python-dotenv
beautifulsoup4
openai
langchain
chromadb

langchain_community
ConversationBufferMemory
tiktoken
pymupdf
openpyxl